Move formatter functions to serializers
