+++ api/serializers/country.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const _ = require('lodash');
+
+function formatCountry(country) {
+  return _.pick(country, ['code', 'name']);
+}
+
+module.exports = {
+  formatCountry,
+};
